Reorganized the metadata caching that was in @memberjunction/aiengine into @memberjunction/ai-engine-base that can be used anywhere and the existing engine derives from the AIEngineBase class and does the server-side only stuff
