+++ packages/AI/Engine/src/index.ts
@@ -1,3 +1,1 @@
 export * from './AIEngine';
-export * from './AIPromptCategoryExtended';
-export * from './AIAgentExtended';
