+++ packages/MJCoreEntities/src/index.ts
@@ -1,7 +1,9 @@
 export * from './generated/entity_subclasses'
 export * from "./custom/UserViewEntity";
 export * from './custom/DashboardEntityExtended';
+export * from './custom/AIAgentExtended';
 export * from './custom/AIModelExtended';
+export * from './custom/AIPromptCategoryExtended';
 export * from './custom/AIPromptExtended';
 export * from './custom/ListDetailEntityExtended';
 export * from './custom/TypeTablesCache';
